feat: ContainFeedDetailCoordinator 파일 분리
