+++ burstcamp/burstcamp/Coordinator/Base/Coordinator.swift
@@ -6,7 +6,6 @@
 //
 
 import Combine
-import SafariServices.SFSafariViewController
 import UIKit
 
 protocol Coordinator: AnyObject {
@@ -33,44 +32,4 @@ protocol NormalCoordinator: Coordinator {
 
 protocol TabBarChildCoordinator: Coordinator {
     func start(viewController: UIViewController)
-}
-
-protocol ContainFeedDetailCoordinator: Coordinator { }
-
-extension ContainFeedDetailCoordinator {
-    func moveToBlogSafari(url: URL) {
-        let safariViewController = SFSafariViewController(url: url)
-        self.navigationController.present(safariViewController, animated: true)
-    }
-
-    func sinkFeedViewController(_ feedDetailViewController: FeedDetailViewController) {
-        feedDetailViewController.coordinatorPublisher
-            .sink { [weak self] event in
-                switch event {
-                case .moveToBlogSafari(let url):
-                    self?.moveToBlogSafari(url: url)
-                }
-            }
-            .store(in: &cancelBag)
-    }
-
-    func prepareFeedDetailViewController(feed: Feed) -> FeedDetailViewController {
-        let feedDetailViewModel = FeedDetailViewModel(feed: feed)
-        let feedScrapViewModel = FeedScrapViewModel(feedUUID: feed.feedUUID)
-        let feedDetailViewController = FeedDetailViewController(
-            feedDetailViewModel: feedDetailViewModel,
-            feedScrapViewModel: feedScrapViewModel
-        )
-        return feedDetailViewController
-    }
-
-    func prepareFeedDetailViewController(feedUUID: String) -> FeedDetailViewController {
-        let feedDetailViewModel = FeedDetailViewModel(feedUUID: feedUUID)
-        let feedScrapViewModel = FeedScrapViewModel(feedUUID: feedUUID)
-        let feedDetailViewController = FeedDetailViewController(
-            feedDetailViewModel: feedDetailViewModel,
-            feedScrapViewModel: feedScrapViewModel
-        )
-        return feedDetailViewController
-    }
 }
